changed toggle to custom togglr && now only requested qouesion have irrelevant answer
