+++ client/src/commons/Toggle/ToggleStyles.ts
@@ -9,11 +9,7 @@ export const useStyles = makeStyles({
     trippleToggle: {
         borderRadius: '50px',
         width: '70px',
-        lineHeight: '1.1rem'
+        lineHeight: '1.1rem',
+        minHeight: '55px'
     },
-    quadrupleToggle: {
-        borderRadius: '50px',
-        width: '80px',
-        lineHeight: '1.1rem'
-    }
 });
