+++ client/src/components/App/Content/InvestigationForm/TabManagement/InteractionsTab/InteractionDialog/InteractionEventForm/InteractionSection/GreenPass/GreenPassQuestioning.tsx
@@ -2,18 +2,19 @@ import React from 'react';
 import {Grid, Typography} from '@material-ui/core';
 import {Controller, useFormContext} from 'react-hook-form';
 
-import Toggle from 'commons/Toggle/Toggle';
 import GreenPassInfo from 'models/GreenPassInfo';
 import FormInput from 'commons/FormInput/FormInput';
+import CustomToggle from 'commons/Toggle/CustomToggle';
 import TripleToggle from 'commons/Toggle/TripleToggle';
 import InteractionEventDialogFields from 'models/enums/InteractionsEventDialogContext/InteractionEventDialogFields';
     
 import useStyles from './GreenPassStyles';
 import useGreenPassQuestioning from './useGreenPassQuestioning';
-import QuadrupleToggle from 'commons/Toggle/QuadrupleToggle';
 
 const greenPassQuestionsTitle = 'הקפדה על נהלים:';
 const finalQuestionsTitle = 'מסקנת החוקר:';
+
+const relevantGreenPassQuestions = [1, 2, 3];
 
 const GreenPassQuestioning = (props :Props) => {
 
@@ -26,27 +27,46 @@ const GreenPassQuestioning = (props :Props) => {
         <>
             <Typography className={classes.title}>{greenPassQuestionsTitle}</Typography>
             <Grid container>
-                { greenPassQuestions.slice(0, -2).map((greenPassQuestion) => {
+                { greenPassQuestions.slice(0, -2).map((greenPassQuestion, index) => {
                     return (
                         <FormInput fieldName={greenPassQuestion.displayName} className={classes.field} isQuestion={true}>
-                            <Controller
-                                name={`${InteractionEventDialogFields.IS_GREEN_PASS}-${greenPassQuestion.id}`}
-                                control={control}
-                                defaultValue={greenPass.hasOwnProperty(greenPassQuestion.id) ? !!greenPass[greenPassQuestion.id] : undefined}
-                                render={(props) => (
-                                    <TripleToggle
-                                        firstOption={greenPassAnswers[0]}
-                                        secondOption={greenPassAnswers[1]}
-                                        thirdOption={greenPassAnswers[greenPassAnswers.length-1]}
-                                        value={props.value}
-                                        onChange={(e, value) => {
-                                            if (value !== null) {
-                                                props.onChange(value)
-                                            }
-                                        }}
-                                    />
-                                )}
-                            />
+                            { relevantGreenPassQuestions.includes(index) ?
+                                <Controller
+                                    name={`${InteractionEventDialogFields.IS_GREEN_PASS}-${greenPassQuestion.id}`}
+                                    control={control}
+                                    defaultValue={greenPass.hasOwnProperty(greenPassQuestion.id) ? greenPass[greenPassQuestion.id] : undefined}
+                                    render={(props) => (
+                                        <CustomToggle
+                                            options={greenPassAnswers.slice(0,2)}
+                                            value={props.value}
+                                            onChange={(e, value) => {
+                                                if (value !== null) {
+                                                    props.onChange(value)
+                                                }
+                                            }}
+                                        />
+                                    )}
+                                />
+                            :
+                                <Controller
+                                    name={`${InteractionEventDialogFields.IS_GREEN_PASS}-${greenPassQuestion.id}`}
+                                    control={control}
+                                    defaultValue={greenPass.hasOwnProperty(greenPassQuestion.id) ? greenPass[greenPassQuestion.id] : undefined}
+                                    render={(props) => (
+                                        <TripleToggle
+                                            firstOption={greenPassAnswers[0]}
+                                            secondOption={greenPassAnswers[1]}
+                                            thirdOption={greenPassAnswers[greenPassAnswers.length-1]}
+                                            value={props.value}
+                                            onChange={(e, value) => {
+                                                if (value !== null) {
+                                                    props.onChange(value)
+                                                }
+                                            }}
+                                        />
+                                    )}
+                                />
+                            }
                         </FormInput>
                     )
                 })}                 
@@ -61,11 +81,8 @@ const GreenPassQuestioning = (props :Props) => {
                                 control={control}
                                 defaultValue={greenPass.hasOwnProperty(finalQuestion.id) ? greenPass[finalQuestion.id] : undefined}
                                 render={(props) => (
-                                    <QuadrupleToggle
-                                        firstOption={greenPassAnswers[greenPassAnswers.length-4]}
-                                        secondOption={greenPassAnswers[greenPassAnswers.length-3]}
-                                        thirdOption={greenPassAnswers[greenPassAnswers.length-2]}
-                                        fourthOption={greenPassAnswers[greenPassAnswers.length-1]}
+                                    <CustomToggle
+                                        options={greenPassAnswers.slice(-4)}
                                         value={props.value}
                                         onChange={(e, value) => {
                                             if (value !== null) {
